이민하 : 사운드 중간커밋

+++ API/EngineBase/EngineRandom.h
@@ -24,7 +24,7 @@ public:
 	{
 		if (_Max < _Min)
 		{
-			float SwapValue = _Min;
+			int SwapValue = _Min;
 			_Min = _Max;
 			_Max = SwapValue;
 		}
